+++ test/OptionsTest.php
@@ -21,6 +21,7 @@ final class OptionsTest extends TestCase
         $this->assertNotNull($options, 'initialize Options');
         $this->assertSame(false, $options->help(), 'help is off.');
     }
+
     public function testH(): void
     {
         $opt = [
@@ -32,6 +33,7 @@ final class OptionsTest extends TestCase
         $this->assertNotNull($options, 'initialize Options');
         $this->assertSame(true, $options->help(), 'help is on.');
     }
+
     public function testHelp(): void
     {
         $opt = [
@@ -43,6 +45,7 @@ final class OptionsTest extends TestCase
         $this->assertNotNull($options, 'initialize Options');
         $this->assertSame(true, $options->help(), 'help is on.');
     }
+
     public function testClassProperties1(): void
     {
         $opt = [
@@ -53,6 +56,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(true, $options->classProperties(), 'classProperties is on.');
     }
+
     public function testClassProperties2(): void
     {
         $opt = [
@@ -63,6 +67,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(false, $options->classProperties(), 'classProperties is off.');
     }
+
     public function testClassProperties3(): void
     {
         $opt = [];
@@ -71,6 +76,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(true, $options->classProperties(), 'classProperties is default on.');
     }
+
     public function testClassMethods1(): void
     {
         $opt = [
@@ -81,6 +87,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(true, $options->classMethods(), 'classMethods is on.');
     }
+
     public function testClassMethods2(): void
     {
         $opt = [
@@ -91,6 +98,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(false, $options->classMethods(), 'classMethods is off.');
     }
+
     public function testClassMethods3(): void
     {
         $opt = [];
@@ -110,6 +118,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(true, $options->classNameSummary(), 'classNameSummary is on.');
     }
+
     public function testClassNameSummary2(): void
     {
         $opt = [
@@ -120,6 +129,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(false, $options->classNameSummary(), 'classNameSummary is off.');
     }
+
     public function testClassNameSummary3(): void
     {
         $opt = [];
@@ -128,7 +138,6 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(true, $options->classNameSummary(), 'classNameSummary is default on.');
     }
-
 
     public function testPhp1(): void
     {
@@ -140,6 +149,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::PHP7, $options->phpVersion(), 'php version is 7.');
     }
+
     public function testPhp2(): void
     {
         $opt = [
@@ -150,6 +160,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::PHP8, $options->phpVersion(), 'php version is 8.');
     }
+
     public function testDiagram1(): void
     {
         $opt = [
@@ -160,6 +171,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::DIAGRAM_CLASS, $options->diagram(), 'diagram is class.');
     }
+
     public function testDiagram2(): void
     {
         $opt = [
@@ -170,6 +182,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::DIAGRAM_PACKAGE, $options->diagram(), 'diagram is package.');
     }
+
     public function testDiagram_division(): void
     {
         $opt = [
@@ -180,6 +193,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::DIAGRAM_DIVISION, $options->diagram(), 'diagram is division.');
     }
+
     public function testDiagram3(): void
     {
         $opt = [];
@@ -188,6 +202,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::DIAGRAM_CLASS, $options->diagram(), 'default diagram is class.');
     }
+
     public function testDiagram4(): void
     {
         $opt = [
@@ -198,6 +213,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(Options::DIAGRAM_JIG, $options->diagram(), 'diagram is jig.');
     }
+
     public function testHeader(): void
     {
         $opt = [
@@ -208,6 +224,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('title PHP Class Diagram', $options->headers()[0], 'specified header.');
     }
+
     public function testMultipleHeaders(): void
     {
         $opt = [
@@ -222,6 +239,7 @@ final class OptionsTest extends TestCase
         $this->assertSame('title PHP Class Diagram', $options->headers()[0], 'specified header. title');
         $this->assertSame('skinparam pageMargin 10', $options->headers()[1], 'specified header. pageMargin');
     }
+
     public function testInclude(): void
     {
         $opt = [
@@ -232,6 +250,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('*.php8', $options->includes()[0], 'specified include.');
     }
+
     public function testMultipleInclude(): void
     {
         $opt = [
@@ -246,6 +265,7 @@ final class OptionsTest extends TestCase
         $this->assertSame('*.php7', $options->includes()[0], 'specified include. php7');
         $this->assertSame('*.php8', $options->includes()[1], 'specified include. php8');
     }
+
     public function testIncludeDefault(): void
     {
         $opt = [];
@@ -254,6 +274,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('*.php', $options->includes()[0], 'default include.');
     }
+
     public function testexclude(): void
     {
         $opt = [
@@ -264,6 +285,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('*Exception.php', $options->excludes()[0], 'specified exclude.');
     }
+
     public function testMultipleExclude(): void
     {
         $opt = [
@@ -278,6 +300,7 @@ final class OptionsTest extends TestCase
         $this->assertSame('*Exception.php', $options->excludes()[0], 'specified exclude. *Exception.php');
         $this->assertSame('config.php', $options->excludes()[1], 'specified exclude. config.php');
     }
+
     public function testExcludeDefault(): void
     {
         $opt = [];
@@ -286,6 +309,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame(0, count($options->excludes()), 'default exclude is empty.');
     }
+
     public function testSvgTopurlDefault(): void
     {
         $opt = [];
@@ -294,6 +318,7 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('*.php', $options->includes()[0], 'default include.');
     }
+
     public function testSvgTopurl(): void
     {
         $opt = [
@@ -304,4 +329,40 @@ final class OptionsTest extends TestCase
 
         $this->assertSame('https://github.com/smeghead/php-class-diagram', $options->svgTopurl(), 'specified svg-topurl.');
     }
+
+    public function testHidePrivateDefault(): void
+    {
+        $opt = [
+        ];
+
+        $options = new Options($opt);
+
+        $this->assertSame(false, $options->hidePrivateProperties(), 'default hide-private-property.');
+        $this->assertSame(false, $options->hidePrivateMethods(), 'default hide-private-methods.');
+    }
+
+    public function testHidePrivateTrue(): void
+    {
+        $opt = [
+            'hide-private' => true,
+        ];
+
+        $options = new Options($opt);
+
+        $this->assertSame(true, $options->hidePrivateProperties(), 'hide-private-property is true.');
+        $this->assertSame(true, $options->hidePrivateMethods(), 'hide-private-methods is true.');
+    }
+
+    public function testHidePrivateMethodsTrue(): void
+    {
+        $opt = [
+            'hide-private-methods' => true,
+        ];
+
+        $options = new Options($opt);
+
+        $this->assertSame(false, $options->hidePrivateProperties(), 'hide-private-property is false.');
+        $this->assertSame(true, $options->hidePrivateMethods(), 'hide-private-methods is true.');
+    }
+
 }