+++ test/fixtures/hide-private/product/Product.php
@@ -1,0 +1,13 @@
+<?php
+namespace hoge\fuga\product;
+
+final class Product {
+    public string $name;
+    private string $memo;
+
+    public function publicFunction(string $param1) {
+    }
+
+    private function privateFunction(string $param1) {
+    }
+}